image added to pane
