+++ src/groupg/controller/AdminLoginController.java
@@ -97,7 +97,7 @@ public class AdminLoginController implements Initializable {
                     System.out.println("Logged into User");
                 }
                 else{
-                    WelcomeScreenController.setPermission(1);
+                    //WelcomeScreenController.setPermission(1);
                     ResourceManager.getInstance().loadFXMLIntoScene("/view/welcomeScreen.fxml", "Admin Main", cancelBtn.getScene());
                     System.out.println("Logged into User");
                 }
